refactor(daco-13): update imports to alias and add new alias
- add components alias
- add pages alias
- update imports to use alias
