+++ apps/ui/src/components/Header.tsx
@@ -21,7 +21,7 @@ import { Button, ConfigProvider, Flex, Image, Layout, Typography } from 'antd';
 
 import PCGL from '@/assets/pcgl-logo-full.png';
 
-import { pcglHeaderTheme } from './providers/ThemeProvider';
+import { pcglHeaderTheme } from '@/components/providers/ThemeProvider';
 
 const { Link } = Typography;
 const { Header } = Layout;
