+++ apps/ui/vite.config.ts
@@ -7,7 +7,9 @@ export default defineConfig({
 	plugins: [react()],
 	resolve: {
 		alias: {
+			'@/components': path.resolve(__dirname, './src/components'),
 			'@/assets': path.resolve(__dirname, './src/assets'),
+			'@/pages': path.resolve(__dirname, './src/pages'),
 		},
 	},
 });
